Split networking from serving in the servers
